requirements.txt
pandas
streamlit
plotly_expres